decouper ray_tracing et prendre l'impact de la lumiere sur la couleur
In minirt.h adding t_color white_light in data and mlx_x mlx_y in data too (before it was in ray_tracing).
In init.c adding the white_light init.
In colors.c adding new functions.
In ray_tracing.c splitting the big function of ray_tracing and dealing with the impact on the color by the ray of light.